requirements.txt
Flask
mysql-connector-python
opencv-python
Pillow
numpy
